@@ -56,26 +55,25 @@
   "devDependencies": {
     "@mcansh/eslint-config": "2.6.1",
     "@svgr/webpack": "5.4.0",
-    "@testing-library/jest-dom": "5.5.0",
-    "@testing-library/react": "10.0.3",
+    "@testing-library/jest-dom": "5.7.0",
+    "@testing-library/react": "10.0.4",
     "@types/body-scroll-lock": "2.6.1",
     "@types/jest": "25.2.1",
     "@types/jest-in-case": "1.0.2",
-    "@types/node": "13.13.1",
+    "@types/node": "13.13.5",
     "@types/nprogress": "0.2.0",
     "@types/prettier": "2.0.0",
     "@types/react": "16.9.34",
-    "@types/react-dom": "16.9.5",
+    "@types/react-dom": "16.9.7",
     "@types/styled-components": "5.1.0",
-    "@typescript-eslint/eslint-plugin": "2.30.0",
-    "@typescript-eslint/parser": "2.30.0",
+    "@typescript-eslint/eslint-plugin": "2.31.0",
+    "@typescript-eslint/parser": "2.31.0",
     "@zeit/next-source-maps": "0.0.4-canary.1",
-    "babel-plugin-macros": "2.8.0",
-    "babel-plugin-preval": "5.0.0",
+    "babel-plugin-import-glob-array": "0.2.0",
     "babel-plugin-styled-components": "1.10.7",
-    "eslint": "6.8.0",
+    "eslint": "7.0.0",
     "eslint-config-airbnb": "18.1.0",
-    "eslint-config-kentcdodds": "14.8.1",
+    "eslint-config-kentcdodds": "14.11.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-import-resolver-typescript": "2.0.0",
     "eslint-plugin-import": "2.20.2",
