@@ -1,8 +1,9 @@
-const fs = require('fs');
-const path = require('path');
+import fs from 'fs';
+import path from 'path';
 
-const { description, homepage } = require('../package.json');
-const posts = require('../data/posts.json');
+import { description, homepage } from '../package.json';
+
+import { posts } from '~/posts';
 
 const OUT_DIR = path.join(process.cwd(), 'public');
 
