@@ -1,6 +1,6 @@
 declare module '*.mdx' {
   import { Post } from '~/components/post-card';
   const MDXComponent: (props) => JSX.Element;
-  export const meta: Post;
+  export const frontMatter: Post;
   export default MDXComponent;
 }
