@@ -8,16 +8,14 @@ const useScrollProgress = () => {
     const bodyHeight = document.body.offsetHeight;
     const windowHeight = window.innerHeight;
     const scrollPercent = scroll / bodyHeight - windowHeight * 100;
-    const maxMinscroll = Math.min(100, Math.max(0, scrollPercent));
-    setScrollProgress(maxMinscroll);
+    const minMaxScroll = Math.min(100, Math.max(0, scrollPercent));
+    setScrollProgress(minMaxScroll);
   }, []);
 
   React.useEffect(() => {
     window.addEventListener('scroll', handleScroll, { passive: true });
     handleScroll();
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
+    return () => window.removeEventListener('scroll', handleScroll);
   }, [handleScroll]);
 
   return scrollProgress;
