@@ -3,7 +3,7 @@ import React from 'react';
 import PostsWrapper from '~/components/posts-wrapper';
 import PostCard, { Post } from '~/components/post-card/index';
 import Header from '~/components/header/index';
-import posts from '~/data/posts.json';
+import { posts } from '~/posts';
 
 const [latest] = posts;
 
