@@ -1,23 +1,9 @@
-import { meta as learningSoftwareDevelopment } from './pages/why-im-learning-software-development.mdx';
-import { meta as html5Progress } from './pages/html5-progress-element.mdx';
-import { meta as objectLifecycle } from './pages/object-lifecycle-cheatsheet.mdx';
-import { meta as javascriptClasses } from './pages/javascript-classes.mdx';
-import { meta as rackKeyConcepts } from './pages/rack-key-concepts.mdx';
-import { meta as lifeAndCode } from './pages/similarities-between-life-and-code.mdx';
-import { meta as timeToHex } from './pages/time-to-hex.mdx';
-import { meta as sinatraProject } from './pages/sinatra-project.mdx';
-import { meta as railsProject } from './pages/rails-project.mdx';
-import { meta as rubyProject } from './pages/ruby-project.mdx';
+// eslint-disable-next-line import/no-unresolved, import/extensions
+import { frontMatter as blogPosts } from './pages/**/*.mdx';
+import { Post } from './components/post-card';
 
-export default [
-  learningSoftwareDevelopment,
-  html5Progress,
-  objectLifecycle,
-  javascriptClasses,
-  rackKeyConcepts,
-  lifeAndCode,
-  timeToHex,
-  sinatraProject,
-  railsProject,
-  rubyProject,
-];
+const posts = ((blogPosts as unknown) as Post[]).sort(
+  (a, b) => new Date(b.date).getTime() - new Date(a.date).getTime()
+);
+
+export { posts };
